Move admin lifecycle functions and rename models to content-types
