+++ packages/core/admin/server/config/index.js
@@ -1,10 +1,6 @@
 'use strict';
 
 module.exports = {
-  functions: {
-    bootstrap: require('./functions/bootstrap'),
-    register: require('./functions/register'),
-  },
   policies: {
     hasPermissions: require('./policies/hasPermissions'),
     isAuthenticatedAdmin: require('./policies/isAuthenticatedAdmin'),
