+++ packages/core/admin/server/destroy.js
@@ -1,0 +1,10 @@
+'use strict';
+
+const { getService } = require('./utils');
+
+module.exports = async () => {
+  const { conditionProvider, actionProvider } = getService('permission');
+
+  await conditionProvider.clear();
+  await actionProvider.clear();
+};
