+++ packages/core/admin/server/register.js
